+++ artifacts/Application/planet9_adaptive_edit/script/neptune.Script/Adaptive/report.js
@@ -447,7 +447,7 @@ const report = {
                     if (s.events && s.events.afterChildSave) s.events.afterChildSave();
 
                     const openedAsSidepanel =
-                        modelAppConfig.oData.settings.navigation.openAs === "S";
+                        modelAppConfig.oData.settings?.navigation?.openAs === "S";
 
                     if (openedAsSidepanel) {
                         // a new record is being created in the sidepanel tab
@@ -491,11 +491,11 @@ const report = {
     },
 
     close: function () {
-        if (
-            oApp.getParent() &&
+        const isDialog = oApp.getParent() &&
             oApp.getParent().getParent() &&
-            oApp.getParent().getParent().close
-        ) {
+            oApp.getParent().getParent().close;
+            
+        if (isDialog) {
             oApp.getParent().getParent().close();
         } else if (
             modelAppConfig.oData.settings.events &&
@@ -508,7 +508,7 @@ const report = {
 
         if (sap.n.Shell && sap.n.Shell.closeAllSidepanelTabs) sap.n.Shell.closeAllSidepanelTabs();
 
-        if (sap.n.HashNavigation && sap.n.HashNavigation.deleteNavItem) {
+        if (!isDialog && sap.n.HashNavigation && sap.n.HashNavigation.deleteNavItem) {
             sap.n.HashNavigation.deleteNavItem();
         }
     },
