+++ artifacts/Application/planet9_adaptive_list/script/neptune.Script/Adaptive/report.js
@@ -604,11 +604,11 @@ const report = {
     },
 
     close: function () {
-        if (
-            oApp.getParent() &&
+        const isDialog = oApp.getParent() &&
             oApp.getParent().getParent() &&
-            oApp.getParent().getParent().close
-        ) {
+            oApp.getParent().getParent().close;
+            
+        if (isDialog) {
             oApp.getParent().getParent().close();
         } else if (
             modelAppConfig.oData.settings.events &&
@@ -621,7 +621,7 @@ const report = {
 
         if (sap.n.Shell && sap.n.Shell.closeAllSidepanelTabs) sap.n.Shell.closeAllSidepanelTabs();
 
-        if (sap.n.HashNavigation && sap.n.HashNavigation.deleteNavItem) {
+        if (!isDialog && sap.n.HashNavigation && sap.n.HashNavigation.deleteNavItem) {
             sap.n.HashNavigation.deleteNavItem();
         }
     },
@@ -1054,7 +1054,7 @@ const report = {
                                 const colData = ctx.getObject();
 
                                 // Sidepanel Lookup Text
-                                if (f._navigation.openAs === "S") {
+                                if (f?._navigation?.openAs === "S") {
                                     const k = f._navigation.dialogTitleField;
                                     const { valueType } = ModelData.FindFirst(
                                         config.settings.fieldsRun,
@@ -1153,7 +1153,7 @@ const report = {
                                 const colData = ctx.getObject();
 
                                 // Sidepanel Lookup Text
-                                if (f._navigation.openAs === "S") {
+                                if (f?._navigation?.openAs === "S") {
                                     const k = f._navigation.dialogTitleField;
                                     const { valueType } = ModelData.FindFirst(
                                         config.settings.fieldsRun,
